+++ libs/libc/stream/lib_lowoutstream.c
@@ -40,8 +40,8 @@
  ****************************************************************************/
 
 static void lowoutstream_putc(FAR struct lib_outstream_s *self, int ch);
-static int lowoutstream_puts(FAR struct lib_outstream_s *self,
-                             FAR const void *buf, int len);
+static ssize_t lowoutstream_puts(FAR struct lib_outstream_s *self,
+                                 FAR const void *buf, size_t len);
 
 /****************************************************************************
  * Public Data
@@ -79,8 +79,8 @@ static void lowoutstream_putc(FAR struct lib_outstream_s *self, int ch)
  * Name: lowoutstream_puts
  ****************************************************************************/
 
-static int lowoutstream_puts(FAR struct lib_outstream_s *self,
-                             FAR const void *buf, int len)
+static ssize_t lowoutstream_puts(FAR struct lib_outstream_s *self,
+                                 FAR const void *buf, size_t len)
 {
   DEBUGASSERT(self);
